requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
channels
channels-redis
celery
redis
drf-spectacular
psycopg2-binary
python-dotenv
